fix(zsh): Fallback to Zoxide on `super-cd`
## Changes Description

Prior to this commit, when using `cd`, I could not change directory to a
known Zoxide directory using just part of its name. For instance, given
this directory structure:

```sh
 ~/
├──  docs/
└──  projects/
    ├──  project1/
    ├──  project2/
    └──  project3/
```

If my `cwd` is `~/docs` and I want to go to `project2`, running:

```sh
cd project2
```

Would result in failure because my `super-cd` command was just using the
_builtin_ `cd` command.

This commit integrates Zoxide with `super-cd` even further by falling
back to standard Zoxide command when there is not directory match for is
arguments.